Move ZonedParser to a separate file
